Revert "COMMERCE-4881 dispatch-service - extract dispatching logic to new class instead of MessageListener"
This reverts commit adb5457f60a9942aa9aca166a7b7e6f187e1e352.
